pyproject.toml
[project]
name = "wep4you-agent-skills"
version = "0.2.0"
description = "Claude Code skills marketplace for Obsidian PKM workflows"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Peter Weiss", email = "github@wep4you.com"}
]
keywords = [
    "claude", "skills", "agent-skills", "obsidian", "pkm", "vault",
    "frontmatter", "validation", "markdown"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Markup :: Markdown",
]

dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "types-PyYAML>=6.0",
]
security = [
    "pip-audit>=2.7.0",
    "bandit>=1.7.8",
    "safety>=3.0.0",
    "semgrep>=1.50.0",
]
all = [
    "wep4you-agent-skills[dev,security]",
]

[project.urls]
Homepage = "https://github.com/wep4you/wep4you-agent-skills"
Repository = "https://github.com/wep4you/wep4you-agent-skills"
Issues = "https://github.com/wep4you/wep4you-agent-skills/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["skills"]

# =============================================================================
# Ruff - Linting & Formatting
# =============================================================================
[tool.ruff]
line-length = 100
target-version = "py310"
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "*.egg-info",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "W",      # pycodestyle warnings
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "S",      # bandit (security)
    "B",      # flake8-bugbear
    "C90",    # mccabe complexity
    "ANN",    # flake8-annotations
    "ASYNC",  # flake8-async
    "PTH",    # flake8-use-pathlib
    "RUF",    # ruff-specific rules
]
ignore = [
    "S101",    # assert usage (needed for tests)
    "ANN204",  # missing return type for __init__
    "ANN201",  # missing return type for public functions (gradual adoption)
    "RUF001",  # ambiguous unicode (emojis allowed)
    "C901",    # complexity (allow for validation scripts)
    "PTH123",  # open() vs Path.open() (both are fine)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "ANN"]

[tool.ruff.lint.mccabe]
max-complexity = 10

# =============================================================================
# MyPy - Type Checking
# =============================================================================
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true
ignore_missing_imports = true

# =============================================================================
# Pytest - Testing
# =============================================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow",
    "security: marks security-related tests",
]

[tool.coverage.run]
source = ["skills", "scripts"]
omit = ["*/tests/*", "*/__pycache__/*"]
branch = true

[tool.coverage.report]
fail_under = 90
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

# =============================================================================
# Bandit - Security Linting
# =============================================================================
[tool.bandit]
exclude_dirs = ["tests", ".venv"]
skips = ["B101"]  # assert_used (needed for tests)

# =============================================================================
# pip-audit - CVE Scanning
# =============================================================================
[tool.pip-audit]
strict = true
vulnerability-service = "osv"
